requirements.txt
Pillow==9.1.1
tflite-model-maker==0.4.1
